backend/requirements.txt
fastapi
uvicorn[standard]
python-multipart
pandas
pydantic
pydantic-settings
python-dotenv
sqlalchemy
psycopg2-binary
asyncpg
alembic
httpx
fpdf2
openpyxl
pytest
pytest-asyncio
RestrictedPython
